Fix bug in velocity conversion
The factor for the velocity conversion from raw to radsec and the other way
round is model-specific and does not depend on the maximum motor speed.
Therefore, a conversion factor is added to the motor data and employed
for all velocity conversions. The set motor velocity and the motor velocity
in the servo state does now correspond with the real servo speed in
radians per second.